+++ account-management-api/src/main/java/uk/gov/di/accountmanagement/entity/mfa/response/ResponseAuthAppMfaDetail.java
@@ -1,4 +1,4 @@
-package uk.gov.di.authentication.shared.entity.mfa.response;
+package uk.gov.di.accountmanagement.entity.mfa.response;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
